server

+++ server_resources/server.c
@@ -4,12 +4,34 @@
 
 char *str = NULL;
 
-void    ft_putchar_fd(char a, int fd)
+void    ft_putstr_fd(char *s, int fd)
 {
-	write(fd, &a, 1);
+	while (s)
+		write(fd, s++, 1);
 }
 
-void	ft_binary_becomes_char(int signum, int bit, siginfo_t *info)
+void	ft_len_to_int(int signum, int i)
+{
+	static int	len = 0;
+
+	if (i < 32)
+	{
+		len = len << 1;
+		if (signum == SIGUSR1)
+			len = len + 1;
+		if (signum == SIGUSR2)
+			len = len + 0;
+	}
+	if (i == 31)
+	{
+		str = malloc(sizeof(char) * (len + 1));
+		if (!str)
+			return ;
+		len = 0;
+	}
+}
+
+void	ft_binary_becomes_char(int signum, int index, int bit, siginfo_t *info, int *i)
 {
 	static int	a = 0;
 
@@ -23,19 +45,38 @@ void	ft_binary_becomes_char(int signum, int bit, siginfo_t *info)
 	}
 	if (bit == 7)
 	{
-		ft_putchar_fd(a, 1);
+		printf("je rentre la1\n");
+		if (str)
+		{
+			printf("je rentre la2\n");
+			str[index] = a;
+			if (a == '\0')
+			{
+				printf("je rentre la3\n");
+				ft_putstr_fd(str, 1);
+				free(str);
+				str = NULL;
+				*i = -1;
+			}
+		}
 		a = 0;
 	}
-	kill(info->si_pid, SIGUSR1);
+	if (bit >= 32)
+	{
+		printf("je rentre ici\n");
+		kill(info->si_pid, SIGUSR1);
+	}
 }
 
 void	ft_signals_handler(int signum, siginfo_t *info, void *name)
 {
+	static int	i = 0;
 	static int	count_bit = 0;
 	(void)name;
 
-	ft_binary_becomes_char(signum, (count_bit % 8), info);
+	ft_binary_becomes_char(signum, ((i - 32) / 8), (count_bit % 8), info, &i);
 	count_bit++;
+	i++;
 }
 
 int main(void)
